UpdateUserController

+++ CheeprToKeepr/Controllers/UsersController.cs
@@ -57,13 +57,22 @@ namespace CheeprToKeepr.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("UserID,FirstName,LastName,AddressLine1,AddressLine2,City,PostalCode,State,Email")] User user)
+        public async Task<IActionResult> Create([Bind("FirstName,LastName,AddressLine1,AddressLine2,City,PostalCode,State,Email")] User user)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(user);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(user);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException)
+            {
+                ModelState.AddModelError("", "Unable to save changes." +
+                    "Try again, and if the problem persists" +
+                    "see your system administrator");
             }
             return View(user);
         }
@@ -87,40 +96,40 @@ namespace CheeprToKeepr.Controllers
         // POST: Users/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("UserID,FirstName,LastName,AddressLine1,AddressLine2,City,PostalCode,State,Email")] User user)
+        public async Task<IActionResult> Edit(int id/*, [Bind("UserID,FirstName,LastName,AddressLine1,AddressLine2,City,PostalCode,State,Email")] User user*/)
         {
-            if (id != user.UserID)
+            if (id == null)
             {
                 return NotFound();
             }
+            var userToUpdate = await _context.Users.FirstOrDefaultAsync(u => u.UserID == id);
 
-            if (ModelState.IsValid)
+            if (await TryUpdateModelAsync<User>(
+                userToUpdate,
+                "",
+                u => u.FirstName, u => u.LastName, u => u.AddressLine1,
+                u => u.AddressLine2, u => u.City, u => u.PostalCode,
+                u => u.State, u => u.Email))
             {
                 try
                 {
-                    _context.Update(user);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!UserExists(user.UserID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Unable To save Changes." +
+                        "Try again, and if the problem persists," +
+                        "sere your system administrator");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(user);
+            return View(userToUpdate);
         }
 
         // GET: Users/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -128,12 +137,19 @@ namespace CheeprToKeepr.Controllers
             }
 
             var user = await _context.Users
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.UserID == id);
             if (user == null)
             {
                 return NotFound();
             }
 
+            if(saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete Failes. Try again, and if the problem persists " +
+                    "see your system administrator.";
+            }
             return View(user);
         }
 
@@ -143,9 +159,21 @@ namespace CheeprToKeepr.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var user = await _context.Users.FindAsync(id);
-            _context.Users.Remove(user);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (user == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+
+                _context.Users.Remove(user);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException)
+            {
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool UserExists(int id)
